new resnet model for audio poisoning
